+++ bitalarm-sdk/src/main/java/com/dwarfeng/bitalarm/sdk/bean/entity/WebInputAlarmTypeIndicator.java
@@ -1,9 +1,11 @@
 package com.dwarfeng.bitalarm.sdk.bean.entity;
 
 import com.alibaba.fastjson.annotation.JSONField;
+import com.dwarfeng.bitalarm.sdk.util.Constraints;
 import com.dwarfeng.bitalarm.stack.bean.entity.AlarmTypeIndicator;
 import com.dwarfeng.subgrade.sdk.bean.key.WebInputStringIdKey;
 import com.dwarfeng.subgrade.stack.bean.Bean;
+import org.hibernate.validator.constraints.Length;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotEmpty;
@@ -18,7 +20,7 @@ import java.util.Objects;
  */
 public class WebInputAlarmTypeIndicator implements Bean {
 
-    private static final long serialVersionUID = -7204862938932762917L;
+    private static final long serialVersionUID = 5809478571833225139L;
 
     public static AlarmTypeIndicator toStackBean(WebInputAlarmTypeIndicator webInputAlarmTypeIndicator) {
         if (Objects.isNull(webInputAlarmTypeIndicator)) {
@@ -26,7 +28,7 @@ public class WebInputAlarmTypeIndicator implements Bean {
         } else {
             return new AlarmTypeIndicator(
                     WebInputStringIdKey.toStackBean(webInputAlarmTypeIndicator.getKey()),
-                    webInputAlarmTypeIndicator.getLabel()
+                    webInputAlarmTypeIndicator.getLabel(), webInputAlarmTypeIndicator.getRemark()
             );
         }
     }
@@ -37,8 +39,14 @@ public class WebInputAlarmTypeIndicator implements Bean {
     private WebInputStringIdKey key;
 
     @JSONField(name = "label")
+    @NotNull
     @NotEmpty
+    @Length(max = Constraints.LENGTH_LABEL)
     private String label;
+
+    @JSONField(name = "remark")
+    @Length(max = Constraints.LENGTH_REMARK)
+    private String remark;
 
     public WebInputAlarmTypeIndicator() {
     }
@@ -59,11 +67,20 @@ public class WebInputAlarmTypeIndicator implements Bean {
         this.label = label;
     }
 
+    public String getRemark() {
+        return remark;
+    }
+
+    public void setRemark(String remark) {
+        this.remark = remark;
+    }
+
     @Override
     public String toString() {
         return "WebInputAlarmTypeIndicator{" +
                 "key=" + key +
                 ", label='" + label + '\'' +
+                ", remark='" + remark + '\'' +
                 '}';
     }
 }