+++ bitalarm-stack/src/main/java/com/dwarfeng/bitalarm/stack/bean/entity/AlarmTypeIndicator.java
@@ -11,17 +11,19 @@ import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
  */
 public class AlarmTypeIndicator implements Entity<StringIdKey> {
 
-    private static final long serialVersionUID = -6629140910520378707L;
+    private static final long serialVersionUID = 7301955764558170245L;
 
     private StringIdKey key;
     private String label;
+    private String remark;
 
     public AlarmTypeIndicator() {
     }
 
-    public AlarmTypeIndicator(StringIdKey key, String label) {
+    public AlarmTypeIndicator(StringIdKey key, String label, String remark) {
         this.key = key;
         this.label = label;
+        this.remark = remark;
     }
 
     @Override
@@ -42,11 +44,20 @@ public class AlarmTypeIndicator implements Entity<StringIdKey> {
         this.label = label;
     }
 
+    public String getRemark() {
+        return remark;
+    }
+
+    public void setRemark(String remark) {
+        this.remark = remark;
+    }
+
     @Override
     public String toString() {
         return "AlarmTypeIndicator{" +
                 "key=" + key +
                 ", label='" + label + '\'' +
+                ", remark='" + remark + '\'' +
                 '}';
     }
 }